refactor(tasks-new): extract task assignee avatars
