+++ src/elements/content-sidebar/activity-feed/task-new/__tests__/AvatarGroup-test.js
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { shallow, mount } from 'enzyme';
 import { TASK_NEW_APPROVED, TASK_NEW_REJECTED, TASK_NEW_NOT_STARTED } from '../../../../../constants';
 
-import Assignees from '../Assignees';
+import AvatarGroup from '../AvatarGroup';
 
 const entries = [
     {
@@ -41,16 +41,16 @@ const assignees = {
 
 const mockGetAvatarUrl = () => Promise.resolve('url.jpg');
 
-describe('elements/content-sidebar/ActivityFeed/task-new/Assignees', () => {
+describe('elements/content-sidebar/ActivityFeed/task-new/AvatarGroup', () => {
     test('should render avatars for each assignee up to maxAvatars', () => {
         const max = 2;
-        const wrapper = mount(<Assignees assignees={assignees} maxAvatars={max} getAvatarUrl={mockGetAvatarUrl} />);
+        const wrapper = mount(<AvatarGroup assignees={assignees} maxAvatars={max} getAvatarUrl={mockGetAvatarUrl} />);
         expect(global.queryAllByTestId(wrapper.render(), 'task-assignment-status')).toHaveLength(max);
     });
 
     test('should show +N overflow when there are more assignees than maxAvatars', () => {
         const max = 2;
-        const wrapper = shallow(<Assignees assignees={assignees} maxAvatars={max} getAvatarUrl={mockGetAvatarUrl} />);
+        const wrapper = shallow(<AvatarGroup assignees={assignees} maxAvatars={max} getAvatarUrl={mockGetAvatarUrl} />);
         const overflowIcon = global.queryAllByTestId(wrapper.dive(), 'task-assignment-overflow');
         expect(overflowIcon).toHaveLength(1);
         expect(overflowIcon.text()).toBe('+1');
@@ -60,7 +60,7 @@ describe('elements/content-sidebar/ActivityFeed/task-new/Assignees', () => {
         const max = 2;
         const assigneesWithPagination = { ...assignees, next_marker: 'abc', limit: 20 };
         const wrapper = shallow(
-            <Assignees assignees={assigneesWithPagination} maxAvatars={max} getAvatarUrl={mockGetAvatarUrl} />,
+            <AvatarGroup assignees={assigneesWithPagination} maxAvatars={max} getAvatarUrl={mockGetAvatarUrl} />,
         );
         const overflowIcon = global.queryAllByTestId(wrapper.dive(), 'task-assignment-overflow');
         expect(overflowIcon).toHaveLength(1);
@@ -69,18 +69,18 @@ describe('elements/content-sidebar/ActivityFeed/task-new/Assignees', () => {
 
     test('should show not show overflow icon when there are fewer assignees than maxAvatars', () => {
         const max = 3;
-        const wrapper = shallow(<Assignees assignees={assignees} maxAvatars={max} getAvatarUrl={mockGetAvatarUrl} />);
+        const wrapper = shallow(<AvatarGroup assignees={assignees} maxAvatars={max} getAvatarUrl={mockGetAvatarUrl} />);
         expect(global.queryAllByTestId(wrapper.render(), 'task-assignment-overflow')).toHaveLength(0);
     });
 
     test('should open assignee list when overflow icon is clicked', () => {
         const max = 2;
-        const wrapper = mount(<Assignees assignees={assignees} maxAvatars={max} getAvatarUrl={mockGetAvatarUrl} />);
+        const wrapper = mount(<AvatarGroup assignees={assignees} maxAvatars={max} getAvatarUrl={mockGetAvatarUrl} />);
         const overflowIcon = global.queryAllByTestId(wrapper, 'task-assignment-overflow').first();
         expect(wrapper.find('Overlay')).toHaveLength(0);
         overflowIcon.simulate('click');
         expect(wrapper.find('Overlay')).toHaveLength(1);
-        const assignmentList = wrapper.find('.bcs-task-assignment-list');
-        expect(assignmentList.find('.bcs-task-assignment-list-item')).toHaveLength(3);
+        const assignmentList = wrapper.find('.bcs-avatar-group-list');
+        expect(assignmentList.find('.bcs-avatar-group-list-item')).toHaveLength(3);
     });
 });
